+++ src/main/java/enterprises/orbital/evexmlapi/eve/ICharacterKillStat.java
@@ -2,9 +2,10 @@ package enterprises.orbital.evexmlapi.eve;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
-@JsonSerialize(as = ICharacterKillStat.class)
+@JsonSerialize(
+    as = ICharacterKillStat.class)
 public interface ICharacterKillStat extends IKillStat {
-  public int getCharacterID();
+  public long getCharacterID();
 
   public String getCharacterName();
 }
