+++ src/main/java/enterprises/orbital/evexmlapi/eve/ICharacterVictoryPointStat.java
@@ -2,9 +2,10 @@ package enterprises.orbital.evexmlapi.eve;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
-@JsonSerialize(as = ICharacterVictoryPointStat.class)
+@JsonSerialize(
+    as = ICharacterVictoryPointStat.class)
 public interface ICharacterVictoryPointStat extends IVictoryPointStat {
-  public int getCharacterID();
+  public long getCharacterID();
 
   public String getCharacterName();
 }
