+++ src/main/java/enterprises/orbital/evexmlapi/eve/IFactionKillStat.java
@@ -2,9 +2,10 @@ package enterprises.orbital.evexmlapi.eve;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
-@JsonSerialize(as = IFactionKillStat.class)
+@JsonSerialize(
+    as = IFactionKillStat.class)
 public interface IFactionKillStat extends IKillStat {
-  public int getFactionID();
+  public long getFactionID();
 
   public String getFactionName();
 }
