+++ src/main/java/enterprises/orbital/evexmlapi/eve/IFactionVictoryPointStat.java
@@ -2,9 +2,10 @@ package enterprises.orbital.evexmlapi.eve;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
-@JsonSerialize(as = IFactionVictoryPointStat.class)
+@JsonSerialize(
+    as = IFactionVictoryPointStat.class)
 public interface IFactionVictoryPointStat extends IVictoryPointStat {
-  public int getFactionID();
+  public long getFactionID();
 
   public String getFactionName();
 }
