+++ src/main/java/enterprises/orbital/evexmlapi/eve/IFactionWar.java
@@ -2,13 +2,14 @@ package enterprises.orbital.evexmlapi.eve;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
-@JsonSerialize(as = IFactionWar.class)
+@JsonSerialize(
+    as = IFactionWar.class)
 public interface IFactionWar {
-  public int getAgainstID();
+  public long getAgainstID();
 
   public String getAgainstName();
 
-  public int getFactionID();
+  public long getFactionID();
 
   public String getFactionName();
 }
